Fix build on clang 4.2.
Thanks Matthieu!

Reviewed-on: http://gerrit.aldebaran.lan:8080/15771
Tested-by: mnottale <mnottale@aldebaran-robotics.com>

+++ qitype/details/functiontypefactory.hxx
@@ -446,7 +446,7 @@ namespace qi
       unsigned long mask = EqFunction<F>::refMask;
       FunctionType* ftype = FunctionTypeEq<MapedF, EqFunPtr>::make(mask, argumentsType, resultType);
 
-      qiLogDebug("qitype.makeGenericFunction") << "bare mask " << EqFunction<F>::refMask;
+      qiLogDebug("qitype.makeGenericFunction") << "bare mask " << (unsigned long)EqFunction<F>::refMask;
       return GenericFunction(ftype, ftype->clone(ftype->initializeStorage(&func)));
     }
